+++ lib/src/server/server.dart
@@ -3,14 +3,14 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:json_api/document.dart';
+import 'package:json_api/src/parser.dart';
 import 'package:json_api/src/server/contracts/controller.dart';
+import 'package:json_api/src/server/contracts/document_builder.dart';
 import 'package:json_api/src/server/contracts/page.dart';
 import 'package:json_api/src/server/contracts/router.dart';
-import 'package:json_api/src/server/contracts/document_builder.dart';
 import 'package:json_api/src/server/standard_document_builder.dart';
 
 part 'server_requests.dart';
-
 part 'server_routes.dart';
 
 class JsonApiServer {