+++ test/unit/document_test.dart
@@ -1,6 +1,3 @@
-import 'dart:convert';
-import 'dart:io';
-
 @TestOn('vm')
 import 'package:json_api/document.dart';
 import 'package:json_matcher/json_matcher.dart';
@@ -18,23 +15,6 @@ void main() {
               'data': {'type': 'foo', 'id': 'bar'}
             }));
       });
-    });
-
-    group('Standard compliance', () {
-      try {
-        test('Can parse the example document', () {
-          // This is a slightly modified example from the JSON:API site
-          // See: https://jsonapi.org/
-          final jsonString =
-              new File('test/unit/example.json').readAsStringSync();
-          final jsonObject = json.decode(jsonString);
-          final doc = Document.parse(jsonObject, ResourceCollectionData.parse);
-
-          expect(doc, encodesToJson(jsonObject));
-        });
-      } catch (e, s) {
-        print(s);
-      }
     });
   });
 }